Made requested changes on PR
