+++ Source/NexusForever.Game.Static/Crafting/CraftingDirection.cs
@@ -1,0 +1,15 @@
+﻿namespace NexusForever.Game.Static.Crafting
+{
+    public enum CraftingDirection
+    {
+        None    = 0x0,
+        N       = 0x1,
+        NE      = 0x2,
+        E       = 0x3,
+        SE      = 0x4,
+        S       = 0x5,
+        SW      = 0x6,
+        W       = 0x7,
+        NW      = 0x8,
+    };
+}
