+++ Source/NexusForever.Game.Static/Crafting/CraftingDiscovery.cs
@@ -1,0 +1,10 @@
+﻿namespace NexusForever.Game.Static.Crafting
+{
+    public enum CraftingDiscovery
+    {
+        Cold    = 0x0,
+        Warm    = 0x1,
+        Hot     = 0x2,
+        Success = 0x3,
+    };
+}
