+++ Source/NexusForever.Game.Static/Crafting/RuneType.cs
@@ -1,0 +1,13 @@
+﻿namespace NexusForever.Game.Static.Crafting
+{
+    public enum RuneType
+    {
+        Air     = 0x7,
+        Fire    = 0xA,
+        Water   = 0x8,
+        Earth   = 0x9,
+        Logic   = 0xB,
+        Life    = 0xC,
+        Fusion  = 0xD,
+    };
+}
