+++ Source/NexusForever.Game.Static/Crafting/TradeskillType.cs
@@ -1,0 +1,19 @@
+﻿namespace NexusForever.Game.Static.Crafting
+{
+    public enum TradeskillType
+    {
+        Weaponsmith = 0x1,
+        Cooking = 0x2,
+        Armorer = 0xC,
+        Mining = 0xD,
+        Outfitter = 0xE,
+        Survivalist = 0xF,
+        Augmentor = 0x10,
+        Architect = 0x11,
+        Relic_Hunter = 0x12,
+        Fishing = 0x13,
+        Farmer = 0x14,
+        Tailor = 0x15,
+        Runecrafting = 0x16,
+    }
+}
