+++ Source/NexusForever.Network.World/Message/Model/Shared/TradeskillInfo.cs
@@ -1,4 +1,5 @@
 ﻿using NexusForever.Network.Message;
+using NexusForever.Game.Static.Crafting;
 
 namespace NexusForever.Network.World.Message.Model.Shared
 {
